third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-iframe-in-main-frame-blocked-expected.txt

CONSOLE ERROR: line 9: [blocked] The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-iframe.html' was loaded over HTTPS, but ran insecure content from 'http://127.0.0.1:8080/security/mixedContent/resources/boring.html': this content should also be loaded over HTTPS.

This test opens a window that loads an insecure iframe. We should not trigger a mixed content callback even though the main frame in the window is HTTPS and is displaying insecure content, because we've set the preference to block this.
